requirements.txt
matplotlib==3.5.0
numpy==1.20.3
Pillow==9.5.0
torch==1.8.0+cu111
torchvision==0.9.0+cu111
tqdm==4.62.3
